Print computed answer when missing reference data
Help solving the problem by outputting the solution we found in skip
message, when there is a puzzle input but no reference output.

This way we avoid manually having to run the CLI with sample data.

Update the skip-tests to allow for skipping the test if no input
was given